Add files via upload
Statistics Canada, one of the most advanced and transparent national statistics agencies in the world,
conducts a variety of surveys. We will focus on a small (1500 row / 20 column) sample of
The Canadian Community Health Survey (CCHS) , including information on height and weight, smoking,
alcohol consumption, general health, socio-demographics, income, and labour force characteristics of the
population.